Crate bp_xcm_bridge

Primitives of the xcm-bridge-hub pallet.

Structs

Bridge

Bridge metadata.

BridgeId

Bridge identifier - used only for communicating with sibling/parent chains in the same consensus.

BridgeLocations

Locations of bridge endpoints at both sides of the bridge.

Deposit

A structure containing information about from whom the deposit is reserved.

Receiver

Receiver metadata.

Enums

BridgeLocationsError

Errors that may happen when we check bridge locations.

BridgeState

Bridge state.

XcmBridgeCall

A minimized version of pallet_xcm_bridge::Call that can be used without a runtime.

Traits

ChannelStatusProvider

Channel status provider that may report whether it is congested or not.

LocalXcmChannelManager

Local XCM channel manager.

Type Aliases

DepositOf

An alias for the bridge deposit of ThisChain.

XcmAsPlainPayload

Encoded XCM blob. We expect the bridge messages pallet to use this blob type for both inbound and outbound payloads.